refactor(toolkit-lib): split resource investigation by resource type
Move investigateEcsService and investigateCustomResource into their own
files (investigate-ecs-service.ts, investigate-custom-resource.ts),
leaving resource-investigation.ts as the dispatcher. Simplify the log
line plumbing: parseLambdaLogEvents now returns string[] (the level is
already combined into each rendered line) and trimToRecentLines operates
on lines rather than raw CloudWatch events.
